% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colTransf}
\alias{colTransf}
\title{Transfers Colors between Meshes}
\usage{
colTransf(M1, M2)
}
\arguments{
\item{M1}{A \code{mesh3d} object to transfer vertex colors from.}

\item{M2}{A \code{mesh3d} object to transfer vertex colors to.}
}
\value{
A decimated \code{mesh3d} object.
}
\description{
This function transfers vertex colors (if any) from one mesh to a one.
}
\details{
This function transfers vertex colors (if any) from the original mesh to the decimated one following
         2 steps: \cr
         - converts of vertex colors from the first mesh \code{M1} to face colors by averaging the colors from
         the 3 vertices of each face, \cr
         - estimates the correspondence among vertices from the second mesh \code{M2} to the first mesh \code{M1}
         by means of the \code{\link[Rvcg]{vcgClostKD}} function).
}
\examples{

}
