% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decimMesh.character}
\alias{decimMesh.character}
\title{Decimates Several Meshes}
\usage{
\method{decimMesh}{character}(M, tarface=15000, sdir=getwd(), patt=".ply", deci.suffix=NULL,
          deci.dir="DecimMesh", \dots)
}
\arguments{
\item{M}{A character value indicating either a mesh filename to decimated within \code{sdir}, or a directory name
within \code{sdir} containing the subdirectory M with the mesh files to decimate.}

\item{tarface}{An integer numerical value for the targetted number of faces for decimation (if superior or equal
to the number of faces in \code{M}, then no decimation is performed).}

\item{sdir}{A character value indicating the directoy name containg eiher the mesh filename \code{M} to decimate,
or the subdirectory name with the mesh files to decimate.}

\item{patt}{A character value within \{\code{".ply"},\code{".stl"}\} indicating the kind of mesh file to open.}

\item{deci.suffix}{A character value indicating the suffix added at the end of the decimated mesh filenames and
before the \code{".ply"} exension. Set to \code{NULL} by default.}

\item{deci.dir}{A character value indicating the name of the subdirectory in \code{M} where the decimated mesh will
be saved. By default, \code{deci.dir} is set to \code{"DecimMesh"}, meaning that a subdirectory
named \code{"DecimMesh"} will be automatically created within \code{M} (if it doesn't exist yet).}

\item{...}{Optional arguments for \code{\link[Rvcg]{vcgQEdecim}}).}
}
\value{
A list of decimated \code{mesh3d} objects.
}
\description{
This function is a wrapper for \code{\link{decimMesh.mesh3d}} (itself a wrapper for
             \code{\link[Rvcg]{vcgQEdecim}}), calling it to treat a list of full mesh files, which will be
             decimated the one after the other. It looks in the directory \code{sdir} for either a mesh filename
             specified by the character value stored in \code{M}, or a subdirectory whose name stored in \code{M}
             that contains the mesh files to decimate. By default, the level of decimation is set by the
             \code{target} argument, but other settings are possible (see \code{\link[Rvcg]{vcgQEdecim}}).
             Decimated mesh are then saved either in the same directory than full meshes, by adding
             a suffix in decimated mesh filenames (to distinguish them from full mesh filenames), or in a separate
             subfolder (contained in the directory for full meshes). Both ply and stl mesh files are supported (
             but decimated meshes will be uniquely saved as ply).
}
\examples{

## Not run:

# 1st example: decimating a mesh file
# A basic call consists in giving the filename of the mesh to decimate (contained in the working directory):
L <- decim("mesh2decimate.ply")

# 2nd example: decimating mesh files in a given directory
# If the working directory contains a subdirectory named "fold" with the full resolution mesh files to decimate, a basic call of the function is:
L <- decim("fold")

# 3rd example: decimating mesh files following a targetted face number percentage (see the help of vcgQEdecim for details):
L <- decim("fold", percent = 0.3)

# 4th example: decimating stl files (contined in the "fold4stl" folder):
L <- decim("fold4stl", patt=".stl")

# 5th example: saving decimating meshes in the same folder as the full meshes by sepcifying their suffix:
L <- decim("fold", deci.suffix="suffix4decimatedMesh", deci.dir=NULL)

## End(Not run)

}
\seealso{
\code{\link{decimMesh.mesh3d}}.
}
