% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decimMesh}
\alias{decimMesh}
\title{Mesh Decimation}
\usage{
decimMesh(M, ...)
}
\arguments{
\item{M}{Either a \code{mesh3d} object (in this case user should refer to \code{\link{decimMesh.mesh3d}}) or a
character value (in this case user should refer to \code{\link{decimMesh.character}})}

\item{...}{Additional arguments for mesh decimation.}
}
\value{
Either \code{mesh3d} object (\code{\link{decimMesh.mesh3d}}), or a list of \code{mesh3d} objects
        (\code{\link{decimMesh.character}})
}
\description{
Generic function for mesh decimation. It invokes 2 particular methods depending on the class of the
             1st argument \code{M}.
}
\details{
For details, user should refer to the 2 methods \code{\link{decimMesh.mesh3d}} (to decimate a single
         \code{mesh3d} object) and \code{\link{decimMesh.character}} (to decimate a set of mesh files).
}
\seealso{
\code{\link{decimMesh.mesh3d}} and \code{\link{decimMesh.character}}.
}
