% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_curves.R
\name{digitCurves.mesh3d}
\alias{digitCurves.mesh3d}
\title{Digitizing 3d curves}
\usage{
digitCurves.mesh3d(
  specFull,
  coords,
  curves,
  npts = 1000,
  nsLds,
  checkMesh = TRUE,
  smoothMesh = NULL,
  GrOpt = setGraphicOptions(),
  verbose = c(TRUE, TRUE),
  spec.name = NULL,
  ...
)
}
\arguments{
\item{specFull}{mesh3d object}

\item{coords}{matrix of 3d coordinates}

\item{...}{additional smoothing parameters (see \code{\link{vcgSmooth}} function)}
}
\value{
a matrix of landmark and semilandmark 3d coordinates plus semiLds and curve attributes
}
\description{
Digitizing 3d curves
}
