% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_fixed_landmarks.R
\name{DigitFixed}
\alias{DigitFixed}
\title{digitFixed}
\usage{
DigitFixed(specFull, specDecim = NULL, decim = 0.5, fixed,
  index = 1:fixed, templateFile = NULL, idxPtsTemplate, orthoplane = TRUE,
  percDist = 0.15, grDev = list(windowRect = rbind(c(0, 50, 830, 904),
  c(840, 50, 1672, 904)), ptSize = 1, spradius = NULL, tcex = 2), ...)
}
\arguments{
\item{specFull}{full resolution mesh3d object}

\item{specDecim}{decimated mesh3d object}

\item{decim}{0-1 number setting the amount of reduction relative to existing face number to decimate the full resolution mesh. Used for multiresolution and only if specDecim is NULL. See vcg:::vcgQEdecim}

\item{fixed}{number of landmarks to digitalize}

\item{index}{Digitalization sequence of landmarks}

\item{templateFile}{template of 3D coordinates. Order of landmarks must be the
same than the one specified in index}

\item{idxPtsTemplate}{Indices of landmarks used to fit the template}

\item{orthoplane}{logical whether or not orthoplane are draw}

\item{percDist}{percentage of distance around the landmark use for zooming}

\item{grDev}{list of graphical parameters with ptSize the point size, windowRect xxxx, spradius and tcex}

\item{...}{additional graphical parameters to be passed to plot (color, alpha, col)}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
cccccc
}

