% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{digitMesh}
\alias{digitMesh}
\title{Mesh Digitization}
\usage{
digitMesh(M, ...)
}
\arguments{
\item{M}{Either a mesh3d object (in this case user should refer to \code{\link{digitMesh.mesh3d}}) or a character
value (in this case user should refer to \code{\link{digitMesh.character}}).}

\item{...}{Additional arguments (all are not optional!) needed for mesh digitization.}
}
\value{
Either a numerical matrix (\code{\link{digitMesh.mesh3d}}) or a numerical array
        (\code{\link{digitMesh.character}}).
}
\description{
Generic function for mesh digitization. It invokes 2 particular methods depending on the class of the
             1st argument \code{M}.
}
\details{
For details, user should refer to the 2 methods \code{\link{digitMesh.mesh3d}} (to digitize a single
         \code{mesh3d} object) and \code{\link{digitMesh.character}} (to digitize a set of mesh files).
}
\seealso{
\code{\link{digitMesh.mesh3d}} and \code{\link{digitMesh.character}}.
}
