% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_3D.R
\name{imputeCoords}
\alias{imputeCoords}
\title{imputeCoords}
\usage{
imputeCoords(A, template)
}
\arguments{
\item{landmark}{matrix with NA or position of landmarks on the specimen}

\item{landmark}{coordinates on the template}
}
\value{
A matrix of imputed landmark coordinates
}
\description{
Function imputes position of landmarks given a full template
and a subset of these landmarks
}
