% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_points_3D.r
\name{PlacePt}
\alias{PlacePt}
\title{PlacePt}
\usage{
PlacePt(x, y, verts, norms, mesh, start)
}
\arguments{
\item{x}{Screen x-coordinate of the clicked point}

\item{y}{Screen y-coordinate of the clicked point}

\item{verts}{XXX}

\item{norms}{XXX}

\item{mesh}{3d mesh}

\item{start}{xxxx}
}
\value{
list with the visible vertices and their index
}
\description{
xxxx
}
\details{
XXX
}

