% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_3D.R
\name{plot.landmark}
\alias{plot.landmark}
\title{plot of one landmark in a 3D scene}
\usage{
plot.landmark(landmark, d1, Sp, Tx, idx_pts, grDev, ...)
}
\arguments{
\item{landmark}{a 3D landmark}

\item{d1}{a rgl scene}

\item{Sp}{xxx}

\item{Tx}{xxxx}

\item{idx_pts}{index of the landmark}

\item{grDev}{xxxx}
}
\description{
Function takes a landmark and plots it onto the surface of a 3D mesh
}

