% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.tps}
\alias{read.tps}
\title{Reads 3D Coordinates from a TPS File}
\usage{
read.tps(file.name, sdir = getwd(), LMheader = "LM3",
         IDheader = "ID", SCheader = "SCALE", k = 3)
}
\arguments{
\item{file.name}{A character value containing the TPS file name (with extension).}

\item{sdir}{A character value indicating the path to the saving directory. \cr
Default: \code{getwd()}.}

\item{LMheader}{A character value specifying the keyword to use in the TPS file to indicate the number of
landmarks. \cr
Default: \code{"LM3"}, for 3D landmarks.}

\item{IDheader}{A character value specifying the keyword to use in the TPS file to indicate the individual
identifier. \cr
Default: \code{"ID"}.}

\item{SCheader}{A character value specifying the keyword to use in the TPS file to indicate the scale
identifier. \cr
Default: \code{"SC"}.}

\item{k}{A numercial value (positive integer) indicating the dimensions of data. \cr
Default: \code{3}.}
}
\value{
A \emph{p*k*n} numerical array (\emph{p}: nb of landmarks, \emph{k}: nb of dimensions, \emph{n}: number
        of individuals).
}
\description{
Reads landmark coordinates stored in a TPS file as generated by \code{\link{save.tps}}. \cr
             \strong{Note}: this is not a generic function to read TPS file, it's only designed for TPS file
                            storing 3D landmark data. For more generic function, see \emph{e.g.}:
                            \code{\link[geomorph]{readland.tps}}, \code{\link[Momocs]{import_tps}} or
                            \code{\link[Morpho]{readallTPS}}.
}
\details{
Given a TPS filename (\code{file.name}) contained in the \code{sdir} directory, this functions extracts 3D
         landmark coordinates (possibly scales them), and stores them into an array (landmarks by lines,
         dimensions=3 by columns, and individual by slices).
}
\examples{

# Generates first two 10*3 random matrices, just for illustration purpose,
# but such as it could be obtained after two mesh digitizations:
M <- matrix(rnorm(30), 10, 3)
N <- matrix(rnorm(30), 10, 3)
# Then, saves them as a TPS file:
save.tps(M, ID = "randomMatrix_M", file.name = "TPSfile4randomMatrix.tps")
save.tps(N, ID = "randomMatrix_N", file.name = "TPSfile4randomMatrix.tps", app=TRUE)
# And finally, read it:
A <- read.tps("TPSfile4randomMatrix.tps")

}
\seealso{
\code{\link{save.tps}}.
}
