% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_points_3D.r
\name{SelectPoints3d}
\alias{SelectPoints3d}
\title{SelectPoints3d}
\usage{
SelectPoints3d(mesh, modify = FALSE, A = NULL, IdxPts = NULL, grDev)
}
\arguments{
\item{mesh}{A 3d mesh as opened by (see \code{\link[Rvcg]{vcgPlyRead}})}

\item{modify}{logical allowing to modify continuously the point by tracking the surface}

\item{A}{Optional matrix that store all landmarks}

\item{IdxPts}{XXX}

\item{grDev}{Graphical parameters (see \code{\link[digit3DLand]{DigitFixed}})}
}
\value{
xxxx vvvvv
}
\description{
Function selects a point on the surface of a 3d mesh
}
\details{
A 3d point is selected using OSX:CMD+mouse, Win:?, Linux:? and
may be optionally moved by not releasing CMD. Final selection is done with ESC
}

