% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setDecimOptions}
\alias{setDecimOptions}
\title{Sets Decimation Options}
\usage{
setDecimOptions(makeDecimation = TRUE, sequential = TRUE,
                tarface = 15000, patt = ".ply")
}
\arguments{
\item{makeDecimation}{A logical value indicating if the decimation should be performed. \cr
Default: \code{TRUE}.}

\item{sequential}{A logical value indicating if the decimation should be performed mesh by mesh during the landmark
digitizing, or in a single pass for all meshes before the digitizing. \cr
Default: \code{TRUE} => decimation mesh by mesh during the digitization process.}

\item{tarface}{A numerical value (positive integer) indicating the targetted number of faces for the decimated
meshes. \cr
Default: \code{15000}.}

\item{patt}{A character value taking values within \{\code{".ply"},\code{".stl"}\} indicating the kind of mesh
files to treat. \cr
Default: \code{".ply"}.}
}
\value{
A list of those parameters.
}
\description{
Set decimation options for \code{\link{digitMesh.character}}.
}
\details{
This function allows user to set several parameters for mesh decimation needed during the call of
         \code{\link{digitMesh.character}}. With no arguments, this function returns the default  settings for
         decimation options. Any of those options can be modified by setting new values for the corresponding
         arguments. Non-filled arguments will be set to default. Ill-filled arguments will return errors or
         warnings.
}
\examples{
# returning default settings:
DeOpt<-setDecimOptions()

# setting for non sequential stl decimations:
DeOpt<-setDecimOptions(sequential = FALSE, patt = ".stl")

}
\seealso{
\code{\link{setFileOptions}}, \code{\link{setGraphicOptions}}, \code{\link{setTemplOptions}}.
}
