% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setFileOptions}
\alias{setFileOptions}
\title{Sets File Options}
\usage{
setFileOptions(M, sdir=getwd(), patt=".ply", deci.suffix=NULL, deci.dir="DecimMesh",
               saveTPS="coord", overwrite=FALSE , append=FALSE)
}
\arguments{
\item{M}{A character value indicating either a mesh filename (with extension) contained in \code{sdir}, or a
subdirectory name (not a path) contained in \code{sdir} in which a set of mesh files should be processed
for digitization.}

\item{sdir}{A character value indicating the path for the directory containing either the mesh to digitize or the
subdirectory containing the mesh files to digitize. \cr
Default: \code{getwd()} => working directory.}

\item{patt}{A character value within \{\code{".ply"},\code{".stl"}\} indicating the kind of mesh file to open. \cr
Default: \code{".ply"}.}

\item{deci.suffix}{A character value indicating the suffix added (or to add) at the end of the decimated mesh
filenames and before the exension. \cr
Default: \code{NULL}.}

\item{deci.dir}{A character value indicating the name of the subdirectory in \code{M} where the decimated mesh will
be saved or extracted. By default, \code{deci.dir} is set to \code{"DecimMesh"}, meaning that a
subdirectory named \code{"DecimMesh"}  will be automatically created within \code{M} (if it doesn't
exist yet).}

\item{saveTPS}{Either a character value indicating the name (with extension) for the TPS filename to export
landmark coordinates, or a \code{FALSE} value if user don't want save coordinate. The TPS file will
be saved within the directory specified by \code{M}. \cr
Default: \code{"coord.tps"}.}

\item{overwrite}{A logical value taking values within \{\code{TRUE},\code{FALSE}\} indicating if the TPS file shoud
be overwritten if the provided filename already exists. \cr
Default: \code{FALSE} => will generate an error if the file aready exists...}

\item{append}{A logical value taking values within \{\code{TRUE},\code{FALSE}\} indicating if the landmark
coordinates should be appended in the TPS file (it could be the case when user need to digitize a set
of meshes in several session but want to store all results in the same file). \cr
Default: \code{FALSE}.}
}
\value{
A list of those parameters plus the information of the directory containing the full mesh(es) (
        \code{full.dir}, as a character value) and the filename(s) of full mesh(es) to treat (\code{full.files}
        as a character vector).
}
\description{
Set file options (opening and saving) for \code{\link{digitMesh.character}}.
}
\details{
This function allows user to set several parameters for mesh file opening and saving as well as TPS export
         needed during the call of \code{\link{digitMesh.character}}. With no arguments, this function returns the
         default settings for file options. Any of those options can be modified by setting new values for the
         corresponding arguments. Non-filled arguments will be set to default. Ill-filled arguments will return
         errors or warnings.

         Depending on the user choice to perform the mesh decimation either during the mesh digitization, or to use
         already decimated meshes, the options linked to the decimated mesh (\code{deci.suffix}, \code{deci.dir})
         will concern either the saving of those meshes, or their opening.

         Additionnally to the provided parameters, this function will return the information of the directory
         containing the full mesh(es) as well as the filename(s) of full mesh(es) to treat.
}
\seealso{
\code{\link{setDecimOptions}}, \code{\link{setGraphicOptions}}, \code{\link{setTemplOptions}}.
}
