% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_points_3D.r
\name{SetPtZoom}
\alias{SetPtZoom}
\title{SetPtZoom}
\usage{
SetPtZoom(dd, specFull, Pt, IdxPts = NULL, orthoplanes, percDist = 0.15,
  modify = FALSE, A = NULL, grDev)
}
\arguments{
\item{dd}{Distance}

\item{specFull}{Full resolution mesh}

\item{Pt}{Numeric vector with the xyz-coordinates of the clicked point}

\item{IdxPts}{xxx}

\item{orthoplanes}{xxx}

\item{percDist}{proportion of the maximu distance between the point and all vertices}

\item{modify}{if the point is modifiable}

\item{A}{Matrix of coordinates of landmarks}

\item{grDev}{Some graphical parameters}
}
\value{
xxx xxxx
}
\description{
Zoom around the selected point on the decimated mesh
}
\details{
Function zoomed on the selected point by opening a new 3d scene with
the full resolution mesh within some distance of the point
}

