% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setTemplOptions}
\alias{setTemplOptions}
\title{Sets Template Options}
\usage{
setTemplOptions(fixed, template = TRUE,
                idxTemplate = 1:4, full.files = FALSE)
}
\arguments{
\item{fixed}{A numerical value (positive integer) indicating the number of landmarks to digitize.}

\item{template}{Possible settings: \cr
- a logical value (\code{TRUE}: the 1st mesh in the directory will serve as template, \code{FALSE}:
  no template will be used); \cr
- a character value indicating the filename (with extension) of the mesh to use for template; \cr
- a numerical \code{fixed*3} matrix containing the template coordinates. \cr
Default: \code{TRUE}.}

\item{idxTemplate}{Numeric vector with positive integers indicating the numbers of landmarks of the template used to
fit it on the mesh, sorted in the order with which they will be digitized on the mesh. For
example, if the landmarks used are the landmarks numbered 10, 12, 17 and 23, and digitized in the
following order on the mesh: 12, 10, 23, 17, \code{idxTemplate} should be set to:
\code{c(12, 10, 23, 17)}. \cr
Default: \code{1:4} => the first four landmarks will serve as reference landmarks.}

\item{full.files}{Either a character vector containing the full mesh filenames to treat (in this case the template
filename given in the \code{template} argument will be ckecked against the filenames provided in
\code{full.files}), or \code{FALSE} (if \code{template} isn't a filename). \cr
Default: \code{FALSE}.}
}
\value{
A list with \code{template}, \code{idxTemplate} plus a logical value \code{makeTempl} indicating if a
        template should be created during the digitization process.
}
\description{
Set options for template during the digitizing process through the call of
             \code{\link{digitMesh.character}}.
}
\details{
This function allows user to set several parameters for the use of a template possibly needed during the
         call of \code{\link{digitMesh.character}}. With no arguments, this function returns the default settings
         for the template options. Any of those options can be modified by setting new values for the corresponding
         arguments. Non-filled arguments will be set to default. Ill-filled arguments will return errors or
         warnings.
}
\examples{
# returning default settings for 30 landmarks:
TeOpt<-setTemplOptions(30)

# setting for using a 30*3 coordinate matrix as a template, and for digitizing landmarks {8,14,29,3}
# (in this order) to fit th template
# 1. generates first a 30*3 random matrix, just for illustration purpose:
M <- matrix(rnorm(90), 30, 3)
TeOpt<-setTemplOptions(30, template = M, idxTemplate = c(8, 14, 29, 3))

}
