% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_3D.R
\name{subset.mesh}
\alias{subset.mesh}
\title{submesh}
\usage{
\method{subset}{mesh}(mesh, subset, select = NULL)
}
\arguments{
\item{mesh}{a mesh3d object}

\item{subset}{expression indicating vertices to keep}

\item{select}{expression indicating fields to select.}
}
\value{
Return of mesh3d object
}
\description{
Extracts a submesh
}
\details{
Function build a submesh from the mesh3d object according to some kept vertices
}

