% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decimMesh.mesh3d}
\alias{decimMesh.mesh3d}
\title{Decimates a Mesh}
\usage{
\method{decimMesh}{mesh3d}(M, tarface = 15000, silent = FALSE, ...)
}
\arguments{
\item{M}{A \code{mesh3d} object.}

\item{tarface}{An integer numerical value for the targetted number of faces for decimation (if superior or equal
to the number of faces in \code{M}, then no decimation is performed).}

\item{silent}{A logical value indicating if console output should be issued once the calls of
\code{\link[Rvcg]{vcgQEdecim}}, \code{\link[Rvcg]{vcgUpdateNormals}} and
\code{\link[Rvcg]{vcgClean}} are done.}

\item{...}{Optional arguments for \code{\link[Rvcg]{vcgQEdecim}}.}
}
\value{
A decimated \code{mesh3d} object.
}
\description{
Wrapper for the \code{\link[Rvcg]{vcgQEdecim}} function making additionnal mesh cleaning needed for
             \code{\link{digitMesh}}.
}
\details{
This function decimates a given mesh \code{M} with \code{\link[Rvcg]{vcgQEdecim}}, then updates mesh
         normals, and removes non manifold faces in the decimated mesh.
}
\examples{

}
\seealso{
\code{\link{decimMesh.character}}.
}
