% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Digit.Fixed.New5.r
\name{DigitFixed}
\alias{DigitFixed}
\title{digitFixed}
\usage{
DigitFixed(specFull, specDecim = NULL, decim = 0.5, fixed,
  templateFile = NULL, idxPtsTemplate, index = 1:fixed, ptSize = 1,
  center = TRUE, wR1 = c(0, 50, 830, 904), wR2 = c(840, 50, 1672, 904),
  orthoplane = TRUE, percDist = 0.15, spradius = NULL, tcex = 2)
}
\arguments{
\item{specFull}{full resolution mesh3d object}

\item{specDecim}{decimated mesh3d object}

\item{decim}{0-1 number setting the amount of reduction relative to existing face number to decimate the full resolution mesh. Used for multiresolution and only if specDecim is NULL. See vcg:::vcgQEdecim}

\item{fixed}{number of landmarks to digitalize}

\item{templateFile}{template of 3D coordinates. Order of landmarks must be the
same than the one specified in index}

\item{idxPtsTemplate}{Indices of landmarks used to fit the template}

\item{index}{Digitalization sequence of landmarks}

\item{ptSize}{point size}

\item{center}{logical whether or not centering is required}

\item{orthoplane}{logical whether or not orthoplane are draw}
}
\description{
XXXX
}
\examples{

}
\author{
Remi Laffont and Nicolas Navarro
}

